+++ src/components/card.js
@@ -1,4 +1,5 @@
 import { openPopup, closePopup } from "./modal.js";
+import { postNewCard, deleteCard, putLike, deleteLike } from "./api.js";
 
 const popupImage = document.querySelector(".popup_type_image");
 const popupImagePic = popupImage.querySelector(".popup__image");
@@ -16,7 +17,9 @@ const createCard = (card) => {
   const cardElementTitle = cardElement.querySelector(".element__title");
   const cardElemenImage = cardElement.querySelector(".element__image");
   const cardElementLike = cardElement.querySelector(".element__like-button");
-  const cardElementLikeCounter = cardElement.querySelector(".element__like-button-counter");
+  const cardElementLikeCounter = cardElement.querySelector(
+    ".element__like-button-counter"
+  );
   const cardElementDelete = cardElement.querySelector(".element__delete");
 
   cardElementTitle.textContent = card.name;
@@ -24,13 +27,13 @@ const createCard = (card) => {
   cardElemenImage.alt = card.name;
   cardElementLikeCounter.textContent = card.likes.length;
 
-  if (card.owner._id === 'f19165c60933c9701244ac4f') {
+  if (card.owner._id === "f19165c60933c9701244ac4f") {
     cardElementDelete.addEventListener("click", () => {
       deleteCard(card._id)
         .then(() => cardElement.remove())
         .catch((err) => {
           console.log(err);
-        });  
+        });
     });
   } else {
     cardElementDelete.remove();
@@ -69,62 +72,5 @@ const addCard = (evt) => {
       console.log(err);
     });
 };
-
-/********************************API***********************************/
-
-const config = {
-  baseUrl: "https://nomoreparties.co/v1/plus-cohort-20",
-  headers: {
-    authorization: "2286b0f6-c117-40dd-b074-20134fb23036",
-    "Content-Type": "application/json",
-  },
-};
-
-const getResOk = (res) => {
-  if (res.ok) {
-    return res.json();
-  }
-  return Promise.reject(`Ошибка: ${res.status}`);
-};
-
-function postNewCard(name, link) {
-  return fetch(`${config.baseUrl}/cards`, {
-    method: "POST",
-    headers: config.headers,
-    body: JSON.stringify({
-      name,
-      link,
-    }),
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
-
-function deleteCard(cardId) {
-  return fetch(`${config.baseUrl}/cards/${cardId}`, {
-    method: "DELETE",
-    headers: config.headers,
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
-
-function putLike(cardId) {
-  return fetch(`${config.baseUrl}/cards/likes/${cardId}`, {
-    method: "PUT",
-    headers: config.headers,
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
-
-function deleteLike(cardId) {
-  return fetch(`${config.baseUrl}/cards/likes/${cardId}`, {
-    method: "DELETE",
-    headers: config.headers,
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
 
 export { createCard, renderInitialCards, addCard, addForm };
