+++ src/index.js
@@ -25,6 +25,28 @@ enableValidation({
   errorClass: "popup__form-input-error_active",
 });
 
+import {
+  getCards,
+  getProfileInfo,
+  patchProfileInfo,
+} from "./components/api.js";
+getCards()
+  .then((cards) => {
+    renderInitialCards(cards);
+  })
+  .catch((err) => {
+    console.log(err);
+  });
+getProfileInfo()
+  .then((data) => {
+    profileName.textContent = data.name;
+    profileStatus.textContent = data.about;
+    profileAvatar.src = data.avatar;
+  })
+  .catch((err) => {
+    console.log(err);
+  });
+
 const editButton = document.querySelector(".profile__edit-button");
 const popupEditProfile = document.querySelector(".popup_type_edit-profile");
 const profile = document.querySelector(".profile__info");
@@ -61,64 +83,3 @@ profileForm.addEventListener("submit", handleProfileFormSubmit);
 
 /********************************API***********************************/
 
-const config = {
-  baseUrl: "https://nomoreparties.co/v1/plus-cohort-20",
-  headers: {
-    authorization: "2286b0f6-c117-40dd-b074-20134fb23036",
-    "Content-Type": "application/json",
-  },
-};
-
-const getResOk = (res) => {
-  if (res.ok) {
-    return res.json();
-  }
-  return Promise.reject(`Ошибка: ${res.status}`);
-};
-
-function getCards() {
-  return fetch(`${config.baseUrl}/cards`, {
-    headers: config.headers,
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
-
-getCards()
-  .then((cards) => {
-    renderInitialCards(cards);
-  })
-  .catch((err) => {
-    console.log(err);
-  });
-
-function getProfileInfo() {
-  return fetch(`${config.baseUrl}/users/me`, {
-    headers: config.headers,
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
-
-getProfileInfo()
-  .then((data) => {
-    profileName.textContent = data.name;
-    profileStatus.textContent = data.about;
-    profileAvatar.src = data.avatar;
-  })
-  .catch((err) => {
-    console.log(err);
-  });
-
-function patchProfileInfo(name, about) {
-  return fetch(`${config.baseUrl}/users/me`, {
-    method: "PATCH",
-    headers: config.headers,
-    body: JSON.stringify({
-      name,
-      about,
-    }),
-  }).then((res) => {
-    return getResOk(res);
-  });
-}
